refactor: change dto
